requirements.txt
python-telegram-bot==20.0
PySocks==1.7.1
